requirements.txt
langchain
langgraph
langchain_groq
llama-index
python-dotenv
requests
beautifulsoup4
lxml